+++ changelog.md
@@ -3,6 +3,7 @@
 * Added pixel-perfect placement mode (Thanks GlobalHive!)
 * Added recipe to duplicate furniture with material
 * Fixed recycling recipe accepting any item
+* Added `requiredCraftingPermissionLevel` to config
 
 # 0.1.3